backend/requirements.txt
africastalking
aiohappyeyeballs
aiohttp
aiohttp-retry
aiosignal
alabaster
alembic
amqp
aniso8601
annotated-types
anyio
attrs
Authlib
awscli
babel
bandit
bcrypt
bidict
billiard
black
bleach
blinker
botocore
brevo-python
cachelib
cachetools
camel-converter
celery
Cerberus
certifi
cffi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
cloudinary
colorama
coverage
cryptography
decorator
defusedxml
Deprecated
dnspython
docutils
dparse
email-validator
et_xmlfile
eventlet
exceptiongroup
execnet
fabric
factory_boy
Faker
filelock
flake8
flasgger
Flask
Flask-Bcrypt
Flask-Caching
flask-cors
Flask-JWT-Extended
Flask-Limiter
Flask-Mail
flask-marshmallow
Flask-Migrate
Flask-RESTful
flask-restx
Flask-Session
Flask-SocketIO
Flask-SQLAlchemy
Flask-Testing
Flask-WTF
flower
freezegun
frozenlist
fsspec
fuzzywuzzy
gitdb
GitPython
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
# --------------------------------------
# Google Auth & OAuth
# --------------------------------------
oauthlib
requests-oauthlib

googleapis-common-protos
greenlet
gunicorn
h11
httpcore
httpie
httplib2
httpx
humanize
idna
imagesize
importlib_resources
iniconfig
invoke
isort
itsdangerous
Jinja2
jmespath
joblib
json-logging
jsonschema
jsonschema-specifications
kombu
Levenshtein
librt
limits
Mako
markdown-it-py
MarkupSafe
marshmallow
marshmallow-sqlalchemy
mccabe
mdurl
meilisearch
mirakuru
mistune
mpmath
msgspec
multidict
mypy
mypy_extensions
networkx
newrelic
nltk
numpy
openpyxl
ordered-set
packaging
pandas
paramiko
passlib
pathspec
pbr
phonenumbers
pillow
platformdirs
pluggy
port-for
prometheus_client
prompt_toolkit
propcache
proto-plus
protobuf
psutil
psycopg
psycopg2
psycopg2-binary
py
py-cpuinfo
pyasn1
pyasn1_modules
pycodestyle
pycparser
pydantic
pydantic_core
pyflakes
Pygments
PyJWT
PyMySQL
PyNaCl
pyotp
pyparsing
PySocks
pytest
pytest-benchmark
pytest-cov
pytest-flask
pytest-html
pytest-metadata
pytest-mock
pytest-postgresql
pytest-xdist
python-dateutil
python-dotenv
python-engineio
python-http-client
python-Levenshtein
python-magic
python-slugify
python-socketio
pytokens
pytz
PyYAML
qrcode
RapidFuzz
redis
referencing
regex
reportlab
requests
requests-toolbelt
responses
retry
rich
rpds-py
rsa
ruamel.yaml
ruamel.yaml.clib
s3transfer
safety
safety-schemas
schema
scikit-learn
scipy
sendgrid
setuptools
shellingham
sib-api-v3-sdk
simple-websocket
six
smmap
sniffio
snowballstemmer
Sphinx
sphinx-rtd-theme
sphinxcontrib-applehelp
sphinxcontrib-devhelp
sphinxcontrib-htmlhelp
sphinxcontrib-jquery
sphinxcontrib-jsmath
sphinxcontrib-qthelp
sphinxcontrib-serializinghtml
SQLAlchemy
starkbank-ecdsa
stevedore
stripe
structlog
sympy
tenacity
text-unidecode
threadpoolctl
tomlkit
tornado
tqdm
twilio
typer
typing-inspection
typing_extensions
tzdata
tzlocal
uritemplate
urllib3
vine
wcwidth
webargs
webencodings
websockets
Werkzeug
wheel
wrapt
wsproto
WTForms
yarl
orjson
Flask-Compress
upstash-redis
redis_cache